requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --extra-index-url=https://download.pytorch.org/whl/nightly/cu130 --pre requirements.in
#
--extra-index-url https://download.pytorch.org/whl/nightly/cu130

accelerate==0.29.3
    # via
    #   -r requirements.in
    #   peft
    #   trl
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via
    #   datasets
    #   fsspec
aiosignal==1.4.0
    # via aiohttp
attrs==25.4.0
    # via aiohttp
bitsandbytes==0.49.0
    # via -r requirements.in
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
cuda-bindings==13.0.3
    # via torch
cuda-pathfinder==1.3.3
    # via cuda-bindings
datasets==2.19.0
    # via
    #   -r requirements.in
    #   trl
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
diskcache==5.6.3
    # via llama-cpp-python
docstring-parser==0.17.0
    # via tyro
filelock==3.20.2
    # via
    #   datasets
    #   huggingface-hub
    #   torch
    #   transformers
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2024.3.1
    # via
    #   datasets
    #   huggingface-hub
    #   torch
gguf==0.17.1
    # via -r requirements.in
huggingface-hub==0.25.2
    # via
    #   -r requirements.in
    #   accelerate
    #   datasets
    #   peft
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   requests
    #   yarl
jinja2==3.1.6
    # via
    #   llama-cpp-python
    #   torch
llama-cpp-python==0.3.16
    # via -r requirements.in
markdown-it-py==4.0.0
    # via
    #   -r requirements.in
    #   rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
networkx==3.6.1
    # via torch
numpy==2.4.0
    # via
    #   -r requirements.in
    #   accelerate
    #   bitsandbytes
    #   datasets
    #   gguf
    #   llama-cpp-python
    #   pandas
    #   peft
    #   transformers
    #   trl
nvidia-cublas==13.1.0.3
    # via
    #   nvidia-cudnn-cu13
    #   nvidia-cusolver
    #   torch
nvidia-cuda-cupti==13.0.85
    # via torch
nvidia-cuda-nvrtc==13.0.88
    # via torch
nvidia-cuda-runtime==13.0.96
    # via torch
nvidia-cudnn-cu13==9.15.1.9
    # via torch
nvidia-cufft==12.0.0.61
    # via torch
nvidia-cufile==1.15.1.6
    # via torch
nvidia-curand==10.4.0.35
    # via torch
nvidia-cusolver==12.0.4.66
    # via torch
nvidia-cusparse==12.6.3.3
    # via
    #   nvidia-cusolver
    #   torch
nvidia-cusparselt-cu13==0.8.0
    # via torch
nvidia-nccl-cu13==2.28.9
    # via torch
nvidia-nvjitlink==13.0.88
    # via
    #   nvidia-cufft
    #   nvidia-cusolver
    #   nvidia-cusparse
    #   torch
nvidia-nvshmem-cu13==3.4.5
    # via torch
nvidia-nvtx==13.0.85
    # via torch
packaging==25.0
    # via
    #   -r requirements.in
    #   accelerate
    #   bitsandbytes
    #   datasets
    #   huggingface-hub
    #   peft
    #   transformers
pandas==3.0.0rc1
    # via datasets
peft==0.10.0
    # via -r requirements.in
pillow==12.1.0
    # via -r requirements.in
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
psutil==7.2.1
    # via
    #   -r requirements.in
    #   accelerate
    #   peft
pyarrow==22.0.0
    # via datasets
pyarrow-hotfix==0.7
    # via datasets
pygments==2.19.2
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via -r requirements.in
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   accelerate
    #   datasets
    #   gguf
    #   huggingface-hub
    #   peft
    #   transformers
regex==2025.11.3
    # via transformers
requests==2.32.5
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
    #   transformers
rich==14.2.0
    # via -r requirements.in
safetensors==0.7.0
    # via
    #   accelerate
    #   peft
    #   transformers
sentencepiece==0.2.1
    # via -r requirements.in
six==1.17.0
    # via python-dateutil
sympy==1.14.0
    # via torch
tokenizers==0.19.1
    # via transformers
torch==2.11.0.dev20260104+cu130
    # via
    #   accelerate
    #   bitsandbytes
    #   peft
    #   trl
tqdm==4.67.1
    # via
    #   datasets
    #   gguf
    #   huggingface-hub
    #   peft
    #   transformers
transformers==4.40.0
    # via
    #   -r requirements.in
    #   peft
    #   trl
triton==3.6.0+git9844da95
    # via torch
trl==0.8.6
    # via -r requirements.in
typeguard==4.4.4
    # via tyro
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   huggingface-hub
    #   llama-cpp-python
    #   torch
    #   typeguard
    #   tyro
tyro==1.0.3
    # via trl
urllib3==2.6.2
    # via requests
xxhash==3.6.0
    # via datasets
yarl==1.22.0
    # via aiohttp
